refactor(StyleOptions): move StyleOptions from Style to its own file
