+++ src/Source/VectorTilesSource.js
@@ -1,5 +1,5 @@
 import { featureFilter } from '@maplibre/maplibre-gl-style-spec';
-import Style from 'Core/Style';
+import StyleOptions from 'Core/StyleOptions';
 import TMSSource from 'Source/TMSSource';
 import URLBuilder from 'Provider/URLBuilder';
 import Fetcher from 'Provider/Fetcher';
@@ -117,7 +117,7 @@ class VectorTilesSource extends TMSSource {
                     if (layer['source-layer'] === undefined) {
                         getPropertiesFromRefLayer(mvtStyle.layers, layer);
                     }
-                    const style = Style.setFromVectorTileLayer(layer, this.sprites, this.symbolToCircle);
+                    const style = StyleOptions.setFromVectorTileLayer(layer, this.sprites, this.symbolToCircle);
                     this.styles[layer.id] = style;
 
                     if (!this.layers[layer['source-layer']]) {
